reverted mapping of version field to integer

+++ src/plugins/kibana/public/discover/saved_searches/_saved_search.js
@@ -36,7 +36,7 @@ define(function (require) {
       hits: 'integer',
       columns: 'string',
       sort: 'string',
-      version: 'long'
+      version: 'integer'
     };
 
     SavedSearch.searchSource = true;
